customize: cast value to intptr_t for pointer usage
Cast a Int64 OCaml value to intptr_t before casting it to a pointer, so
this avoids warnings about casting an int to pointer.

Fixes commit 80a13b7216340b2bee437c39db2375b61804e133.

(cherry picked from commit f5821fe6b798b3fc14db4ad82b5af52f9fd13e69)

+++ customize/perl_edit-c.c
@@ -37,7 +37,7 @@ virt_customize_edit_file_perl (value verbosev, value gv, value gpv,
 {
   CAMLparam5 (verbosev, gv, gpv, filev, exprv);
   int r;
-  guestfs_h *g = (guestfs_h *) Int64_val (gpv);
+  guestfs_h *g = (guestfs_h *) (intptr_t) Int64_val (gpv);
 
   r = edit_file_perl (g, String_val (filev), String_val (exprv), NULL,
                       Bool_val (verbosev));
